--- module-core/src/main/java/kitchenpos/core/domain/Name.java
@@ -1,38 +1,0 @@
-package kitchenpos.core.domain;
-
-import javax.persistence.Column;
-import javax.persistence.Embeddable;
-import java.util.Objects;
-
-import static java.util.Objects.requireNonNull;
-
-@Embeddable
-public class Name {
-
-    @Column(nullable = false)
-    private String name;
-
-    public Name(String name) {
-        this.name = requireNonNull(name, "name");
-    }
-
-    protected Name() {
-    }
-
-    public String getValue() {
-        return name;
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (o == null || getClass() != o.getClass()) return false;
-        Name that = (Name) o;
-        return Objects.equals(name, that.name);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(name);
-    }
-}
