+++ module-menu/src/main/java/kitchenpos/menu/dto/MenuProductResponse.java
@@ -1,0 +1,49 @@
+package kitchenpos.menu.dto;
+
+import kitchenpos.menu.domain.MenuProduct;
+import java.util.List;
+import java.util.stream.Collectors;
+
+public class MenuProductResponse {
+    private Long seq;
+    private Long menuId;
+    private Long productId;
+    private long quantity;
+
+    protected MenuProductResponse(Long seq, Long menuId, Long productId, long quantity) {
+        this.seq = seq;
+        this.menuId = menuId;
+        this.productId = productId;
+        this.quantity = quantity;
+    }
+
+    protected MenuProductResponse() {
+    }
+
+    public static List<MenuProductResponse> of(List<MenuProduct> menuProducts) {
+        return menuProducts.stream()
+                           .map(MenuProductResponse::of)
+                           .collect(Collectors.toList());
+    }
+
+    public static MenuProductResponse of(MenuProduct menuProduct) {
+        return new MenuProductResponse(menuProduct.getSeq(), menuProduct.getMenu().getId(),
+                                       menuProduct.getProductId(), menuProduct.getQuantity().getValue());
+    }
+
+    public Long getSeq() {
+        return seq;
+    }
+
+    public Long getMenuId() {
+        return menuId;
+    }
+
+    public Long getProductId() {
+        return productId;
+    }
+
+    public long getQuantity() {
+        return quantity;
+    }
+}
