+++ module-menu/src/main/java/kitchenpos/menu/dto/MenuRequest.java
@@ -1,0 +1,46 @@
+package kitchenpos.menu.dto;
+
+import kitchenpos.menu.domain.Menu;
+import kitchenpos.menu.domain.MenuProduct;
+import java.math.BigDecimal;
+import java.util.List;
+import java.util.stream.Collectors;
+
+public class MenuRequest {
+    private String name;
+    private BigDecimal price;
+    private Long menuGroupId;
+    private List<MenuProductRequest> menuProducts;
+
+    public MenuRequest(String name, BigDecimal price, Long menuGroupId, List<MenuProductRequest> menuProducts) {
+        this.name = name;
+        this.price = price;
+        this.menuGroupId = menuGroupId;
+        this.menuProducts = menuProducts;
+    }
+
+    protected MenuRequest() {
+    }
+
+    public Menu toMenu() {
+        return new Menu(name, price, menuGroupId);
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public BigDecimal getPrice() {
+        return price;
+    }
+
+    public Long getMenuGroupId() {
+        return menuGroupId;
+    }
+
+    public List<MenuProduct> getMenuProducts() {
+        return menuProducts.stream()
+                           .map(MenuProductRequest::toMenuProduct)
+                           .collect(Collectors.toList());
+    }
+}
