+++ module-menu/src/test/java/kitchenpos/MenuApplication.java
@@ -1,0 +1,12 @@
+package kitchenpos;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class MenuApplication {
+
+    @Test
+    public void contextLoads() {
+    }
+}
