+++ module-menugroup/src/main/java/kitchenpos/menugroup/domain/MenuGroupRepository.java
@@ -1,0 +1,8 @@
+package kitchenpos.menugroup.domain;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface MenuGroupRepository extends JpaRepository<MenuGroup, Long> {
+
+    boolean existsById(Long id);
+}
