+++ module-menugroup/src/main/java/kitchenpos/menugroup/dto/MenuGroupResponse.java
@@ -1,0 +1,36 @@
+package kitchenpos.menugroup.dto;
+
+import kitchenpos.menugroup.domain.MenuGroup;
+import java.util.List;
+import java.util.stream.Collectors;
+
+public class MenuGroupResponse {
+    private Long id;
+    private String name;
+
+    protected MenuGroupResponse(Long id, String name) {
+        this.id = id;
+        this.name = name;
+    }
+
+    protected MenuGroupResponse() {
+    }
+
+    public static MenuGroupResponse of(MenuGroup menuGroup) {
+        return new MenuGroupResponse(menuGroup.getId(), menuGroup.getName());
+    }
+
+    public static List<MenuGroupResponse> of(List<MenuGroup> menuGroups) {
+        return menuGroups.stream()
+                         .map(MenuGroupResponse::of)
+                         .collect(Collectors.toList());
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public String getName() {
+        return name;
+    }
+}
