+++ module-menugroup/src/main/java/kitchenpos/menugroup/exception/NotFoundMenuGroupException.java
@@ -1,0 +1,8 @@
+package kitchenpos.menugroup.exception;
+
+public class NotFoundMenuGroupException extends RuntimeException {
+
+    public NotFoundMenuGroupException() {
+        super("메뉴 그룹을 찾을 수 없습니다.");
+    }
+}
