+++ module-menugroup/src/test/java/kitchenpos/MenuGroupApplicationTests.java
@@ -1,0 +1,12 @@
+package kitchenpos;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class MenuGroupApplicationTests {
+
+    @Test
+    public void contextLoads() {
+    }
+}
