+++ module-menugroup/src/test/java/kitchenpos/menugroup/application/MenuGroupServiceTest.java
@@ -1,0 +1,30 @@
+package kitchenpos.menugroup.application;
+
+import kitchenpos.ServiceTest;
+import kitchenpos.menugroup.dto.MenuGroupRequest;
+import kitchenpos.menugroup.dto.MenuGroupResponse;
+import kitchenpos.menugroup.application.MenuGroupService;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.junit.jupiter.api.Assertions.assertAll;
+
+class MenuGroupServiceTest extends ServiceTest {
+
+    @Autowired
+    private MenuGroupService service;
+
+    @DisplayName("메뉴 그룹을 생성한다.")
+    @Test
+    void create() {
+        MenuGroupRequest request = new MenuGroupRequest("신메뉴");
+        MenuGroupResponse response = service.create(request);
+
+        assertAll(
+                () -> assertThat(response.getId()).isNotNull(),
+                () -> assertThat(response.getName()).isEqualTo("신메뉴")
+        );
+    }
+}
