+++ module-tablegroup/build.gradle
@@ -1,0 +1,4 @@
+dependencies {
+    implementation project(':module-table')
+    testFixturesApi(testFixtures(project(':module-table')))
+}
