--- src/main/java/kitchenpos/menu/domain/MenuGroup.java
@@ -1,44 +1,0 @@
-package kitchenpos.menu.domain;
-
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-
-import static java.util.Objects.requireNonNull;
-
-@Entity
-public class MenuGroup {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-    private String name;
-
-    protected MenuGroup() {
-    }
-
-    public MenuGroup(String name) {
-        requireNonNull(name, "name");
-        this.name = name;
-    }
-
-    private void validate() {
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(final Long id) {
-        this.id = id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(final String name) {
-        this.name = name;
-    }
-
-}
