--- src/main/java/kitchenpos/menu/dto/MenuGroupRequest.java
@@ -1,22 +1,0 @@
-package kitchenpos.menu.dto;
-
-import kitchenpos.menu.domain.MenuGroup;
-
-public class MenuGroupRequest {
-    private String name;
-
-    public MenuGroupRequest(String name) {
-        this.name = name;
-    }
-
-    protected MenuGroupRequest() {
-    }
-
-    public MenuGroup toMenuGroup() {
-        return new MenuGroup(name);
-    }
-
-    public String getName() {
-        return name;
-    }
-}
