--- src/main/java/kitchenpos/menu/dto/MenuRequest.java
@@ -1,43 +1,0 @@
-package kitchenpos.menu.dto;
-
-import kitchenpos.menu.domain.Menu;
-
-import java.math.BigDecimal;
-import java.util.List;
-
-public class MenuRequest {
-    private String name;
-    private BigDecimal price;
-    private Long menuGroupId;
-    private List<MenuProductRequest> menuProducts;
-
-    public MenuRequest(String name, BigDecimal price, Long menuGroupId, List<MenuProductRequest> menuProducts) {
-        this.name = name;
-        this.price = price;
-        this.menuGroupId = menuGroupId;
-        this.menuProducts = menuProducts;
-    }
-
-    protected MenuRequest() {
-    }
-
-    public Menu toMenu() {
-        return new Menu(name, price, menuGroupId);
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public BigDecimal getPrice() {
-        return price;
-    }
-
-    public Long getMenuGroupId() {
-        return menuGroupId;
-    }
-
-    public List<MenuProductRequest> getMenuProducts() {
-        return menuProducts;
-    }
-}
