--- src/main/java/kitchenpos/menu/dto/MenuResponse.java
@@ -1,58 +1,0 @@
-package kitchenpos.menu.dto;
-
-import kitchenpos.menu.domain.Menu;
-
-import java.math.BigDecimal;
-import java.util.List;
-import java.util.stream.Collectors;
-
-public class MenuResponse {
-    private Long id;
-    private String name;
-    private BigDecimal price;
-    private Long menuGroupId;
-    private List<MenuProductResponse> menuProducts;
-
-    protected MenuResponse(Long id, String name, BigDecimal price, Long menuGroupId,
-                           List<MenuProductResponse> menuProducts) {
-        this.id = id;
-        this.name = name;
-        this.price = price;
-        this.menuGroupId = menuGroupId;
-        this.menuProducts = menuProducts;
-    }
-
-    protected MenuResponse() {
-    }
-
-    public static List<MenuResponse> of(List<Menu> menus) {
-        return menus.stream()
-                .map(MenuResponse::of)
-                .collect(Collectors.toList());
-    }
-
-    public static MenuResponse of(Menu menu) {
-        return new MenuResponse(menu.getId(), menu.getName(), menu.getPrice().getValue(),
-                menu.getMenuGroupId(), MenuProductResponse.of(menu.getMenuProducts()));
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public BigDecimal getPrice() {
-        return price;
-    }
-
-    public Long getMenuGroupId() {
-        return menuGroupId;
-    }
-
-    public List<MenuProductResponse> getMenuProducts() {
-        return menuProducts;
-    }
-}
