--- src/main/java/kitchenpos/menu/exception/NotFoundMenuGroupException.java
@@ -1,8 +1,0 @@
-package kitchenpos.menu.exception;
-
-public class NotFoundMenuGroupException extends RuntimeException {
-
-    public NotFoundMenuGroupException() {
-        super("메뉴 그룹을 찾을 수 없습니다.");
-    }
-}
